Rearranging and fixing for file names
